feat: Add error handling in AI recommender
